# openspec/specs/openapi.yaml
openapi: 3.1.0
info:
  title: BeBudget API
  version: '1'
  description: 'Operational header behavior: responses include `X-Request-Id`.

    If request provides `X-Request-Id`, the API echoes it; otherwise it generates one.

    Cross-site CORS policy for browser clients: API uses explicit origin allowlist from configuration
    (`BEBUDGET_CORS_ORIGINS`, default `http://localhost:5173`), `Access-Control-Allow-Credentials:
    true`, allowed methods `GET,POST,PATCH,DELETE,OPTIONS`, allowed headers `Authorization,Content-Type,Accept,X-Request-Id`,
    and exposed headers `X-Request-Id,Retry-After`.

    Baseline API security headers: `X-Content-Type-Options: nosniff`, `Referrer-Policy: no-referrer`,
    `Cross-Origin-Opener-Policy: same-origin`, and `Content-Security-Policy: default-src ''none''; frame-ancestors ''none''; base-uri ''none''`.

    '
servers:
- url: /api
security:
- BearerAuth: []
paths:
...
  /categories:
    get:
      summary: List categories
      description: 'Archived policy: categories are excluded by default and included only when `include_archived=true`.


        Cursor pagination uses an opaque `base64url(JSON)` token.

        Ordering is deterministic by `created_at DESC, id DESC`, and `next_cursor` is derived from that same sort key.

        For stable datasets, paging is deterministic with no overlaps or skips; terminal page returns `next_cursor: null`.

        Under concurrent mutations, pagination is best-effort and does not guarantee snapshot semantics.

        '
      parameters:
      - name: include_archived
        in: query
        required: false
        schema:
          type: boolean
          default: false
        description: Include archived categories when true. Defaults to false (active categories only).
      - name: type
        in: query
        required: false
        schema:
          type: string
          enum:
          - income
          - expense
      - name: cursor
        in: query
        required: false
        description: Opaque `base64url(JSON)` token built from `created_at` and `id` of the last item in the previous page.
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
          maximum: 100
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/vnd.bebudget.v1+json:
              schema:
                type: object
                required:
                - items
                - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
                  next_cursor:
                    type:
                    - string
                    - 'null'
              example:
                items:
                - &id002
                  id: 0dab8db1-dcb8-44a2-bccd-d2efce11c7c9
                  name: Groceries
                  type: expense
                  note: Supermarket and food
                  archived_at: null
                next_cursor: null
        '400':
          description: Invalid cursor
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem400InvalidCursor'
        '401':
          description: Unauthorized
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem401'
        '406':
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem406'
    post:
      summary: Create category
      parameters:
      - name: on_conflict
        in: query
        required: false
        schema:
          type: string
          enum:
          - error
          - use_existing
          default: error
        description: Conflict strategy. `error` keeps `409` duplicate behavior; `use_existing` returns existing/restored category.
      requestBody:
        required: true
        content:
          application/vnd.bebudget.v1+json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '200':
          description: Existing category reused or restored
          content:
            application/vnd.bebudget.v1+json:
              schema:
                $ref: '#/components/schemas/Category'
        '201':
          description: Created
          content:
            application/vnd.bebudget.v1+json:
              schema:
                $ref: '#/components/schemas/Category'
              example:
                id: 0dab8db1-dcb8-44a2-bccd-d2efce11c7c9
                name: Groceries
                type: expense
                note: Supermarket and food
                archived_at: null
        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem400InvalidRequest'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem401'
        '409':
          description: Category name already exists for this type
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem409Conflict'
        '406':
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem406'
  /categories/{category_id}:
    parameters:
    - name: category_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    get:
      summary: Get category
      responses:
        '200':
          description: OK
          content:
            application/vnd.bebudget.v1+json:
              schema:
                $ref: '#/components/schemas/Category'
              example:
                id: 0dab8db1-dcb8-44a2-bccd-d2efce11c7c9
                name: Groceries
                type: expense
                note: Supermarket and food
                archived_at: null
        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem400InvalidRequest'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem401'
        '403':
          description: Forbidden (resource is not owned by authenticated user)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem403'
        '406':
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem406'
      description: Returns an owned category regardless of archive state (active or archived).
    patch:
      summary: Update category
      requestBody:
        required: true
        content:
          application/vnd.bebudget.v1+json:
            schema:
              $ref: '#/components/schemas/CategoryUpdate'
      responses:
        '200':
          description: OK
          content:
            application/vnd.bebudget.v1+json:
              schema:
                $ref: '#/components/schemas/Category'
              example:
                id: 0dab8db1-dcb8-44a2-bccd-d2efce11c7c9
                name: Groceries
                type: expense
                note: Supermarket and food
                archived_at: null
        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem400InvalidRequest'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem401'
        '403':
          description: Forbidden (resource is not owned by authenticated user)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem403'
        '409':
          description: Category name already exists for this type
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem409Conflict'
        '406':
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem406'
    delete:
      summary: Archive category
      responses:
        '204':
          description: Archived (soft-delete). No Content
        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem400InvalidRequest'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem401'
        '403':
          description: Forbidden (resource is not owned by authenticated user)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem403'
        '406':
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                canonical:
                  $ref: '#/components/examples/Problem406'
...
components:
  headers:
    X-Request-Id:
      description: Correlation id for tracing requests. Propagated when provided by client, generated otherwise.
      schema:
        type: string
...
  schemas:
    ProblemDetails:
      type: object
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
      examples:
      - type: https://api.bebudget.dev/problems/invalid-date-range
        title: Invalid date range
        status: 400
        detail: '`from` must be less than or equal to `to`.'
      - type: https://api.bebudget.dev/problems/unauthorized
        title: Unauthorized
        status: 401
        detail: Missing or invalid bearer token.
      - type: https://api.bebudget.dev/problems/forbidden
        title: Forbidden
        status: 403
        detail: Resource is not owned by authenticated user.
      - type: https://api.bebudget.dev/problems/not-acceptable
        title: Not Acceptable
        status: 406
        detail: Accept header must allow a supported media type.
      - type: https://api.bebudget.dev/problems/account-archived
        title: Account is archived
        status: 409
        detail: Transactions cannot be created on archived accounts.
      - type: https://api.bebudget.dev/problems/rate-limited
        title: Too Many Requests
        status: 429
        detail: Too many requests. Retry after the interval in Retry-After.
...
    Category:
      type: object
      required:
      - id
      - name
      - type
      - archived_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum:
          - income
          - expense
        note:
          type: string
        archived_at:
          type:
          - string
          - 'null'
          format: date-time
      example: *id002
    CategoryCreate:
      type: object
      required:
      - name
      - type
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum:
          - income
          - expense
        note:
          type: string
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        note:
          type: string
        archived_at:
          type:
          - string
          - 'null'
          format: date-time
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  examples:
    Problem400InvalidRequest:
      summary: Generic 400 invalid request
      value:
        type: about:blank
        title: Invalid request
        status: 400
        detail: Request payload or parameters are invalid.
    Problem400InvalidCursor:
      summary: Canonical 400 invalid cursor
      value:
        type: https://api.bebudget.dev/problems/invalid-cursor
        title: Invalid cursor
        status: 400
        detail: Cursor token is malformed or unsupported.
...
    Problem401:
      summary: Canonical 401 error
      value:
        type: https://api.bebudget.dev/problems/unauthorized
        title: Unauthorized
        status: 401
        detail: Missing or invalid bearer token.
    Problem403:
      summary: Canonical 403 error
      value:
        type: https://api.bebudget.dev/problems/forbidden
        title: Forbidden
        status: 403
        detail: Resource is not owned by authenticated user.
...
    Problem406:
      summary: Canonical 406 error
      value:
        type: https://api.bebudget.dev/problems/not-acceptable
        title: Not Acceptable
        status: 406
        detail: Accept header must allow a supported media type.
    Problem409Conflict:
      summary: Generic 409 conflict
      value:
        type: about:blank
        title: Conflict
        status: 409
        detail: Resource conflict with current state.
...